+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 
 ### Changed
 
+- [#6471](https://github.com/ChainSafe/forest/pull/6471) Moved `forest-tool state` subcommand to `forest-dev`.
+
 ### Removed
 
 ### Fixed